requirements.txt
Flask
psycopg2
SQLAlchemy
gunicorn
python-dotenv
requests
oauthlib